vendor_pricing/requirements.txt
playwright==1.44.0
gspread==6.1.2
google-auth==2.29.0
click==8.1.7
rich==13.7.1
flask==3.0.3
python-dotenv==1.0.1
pytest==8.2.0
